Fixed Model Clipping
The engineer model had two models for his right hand that were clipping into eachother so I turned the robot version off to resolve the clipping. If the model is used in other scenes this change will not apply until you disable the "robotarm_bodygroup" child under the engineer model.