Rework of Player Stats
Guys this was a big one. I have reworked every script that had to do with the current player statistics (health, and speed) I had to link some assembly definitions (which I relearned the hard way is how to use a class from another folder) and I had to replace a lot of references to walk speed in order to replace them with the new method of getting the speed by using the getter function in PlayerStats. If your stuff doesn't work now please tell me because it all seems to function perfectly fine to me! I'll be starting on implementing game manager now that health is actually stored on the player correctly.